Added a dedicated MappingException extending the Common one.
