+++ lib/Doctrine/ODM/MongoDB/DocumentRepository.php
@@ -21,6 +21,7 @@
 namespace Doctrine\ODM\MongoDB;
 
 use Doctrine\Common\Persistence\ObjectRepository;
+use Doctrine\ODM\MongoDB\Mapping\MappingException;
 
 /**
  * An DocumentRepository serves as a repository for documents with generic as well as
@@ -53,7 +54,7 @@ class DocumentRepository implements ObjectRepository
     protected $uow;
 
     /**
-     * @var Doctrine\ODM\MongoDB\Mapping\ClassMetadata
+     * @var \Doctrine\ODM\MongoDB\Mapping\ClassMetadata
      */
     protected $class;
 
@@ -62,7 +63,7 @@ class DocumentRepository implements ObjectRepository
      *
      * @param DocumentManager $dm The DocumentManager to use.
      * @param UnitOfWork $uow The UnitOfWork to use.
-     * @param ClassMetadata $classMetadata The class descriptor.
+     * @param Mapping\ClassMetadata $classMetadata The class descriptor.
      */
     public function __construct(DocumentManager $dm, UnitOfWork $uow, Mapping\ClassMetadata $class)
     {
@@ -108,12 +109,12 @@ class DocumentRepository implements ObjectRepository
             list($identifierFieldName) = $this->class->getIdentifierFieldNames();
 
             if (!isset($id[$identifierFieldName])) {
-                throw MongoDBException::missingIdentifierField($this->documentName, $identifierFieldName);
+                throw MappingException::missingIdentifierField($this->documentName, $identifierFieldName);
             }
 
             $id = $id[$identifierFieldName];
         }
-        
+
         // Check identity map first
         if ($document = $this->uow->tryGetById($id, $this->class->rootDocumentName)) {
             if ($lockMode != LockMode::NONE) {
