+++ tests/Doctrine/ODM/MongoDB/Tests/Mapping/ClassMetadataTest.php
@@ -97,7 +97,7 @@ class ClassMetadataTest extends \Doctrine\ODM\MongoDB\Tests\BaseTest
 
         $this->assertEquals("DoctrineGlobal_User", $cm->fieldMappings['author']['targetDocument']);
     }
-    
+
     public function testMapManyToManyJoinTableDefaults()
     {
         $cm = new ClassMetadata('Documents\CmsUser');
@@ -106,7 +106,7 @@ class ClassMetadataTest extends \Doctrine\ODM\MongoDB\Tests\BaseTest
             'fieldName' => 'groups',
             'targetDocument' => 'CmsGroup'
         ));
-        
+
         $assoc = $cm->fieldMappings['groups'];
         $this->assertTrue(is_array($assoc));
     }
@@ -145,7 +145,7 @@ class ClassMetadataTest extends \Doctrine\ODM\MongoDB\Tests\BaseTest
         $a2 = array('reference' => true, 'type' => 'many', 'fieldName' => 'foo', 'targetDocument' => 'stdClass');
 
         $cm->mapField($a1);
-        $this->setExpectedException('Doctrine\ODM\MongoDB\MongoDBException');
+        $this->setExpectedException('Doctrine\ODM\MongoDB\Mapping\MappingException');
         $cm->mapField($a2);
     }
 
@@ -154,7 +154,7 @@ class ClassMetadataTest extends \Doctrine\ODM\MongoDB\Tests\BaseTest
         $cm = new ClassMetadata('Documents\CmsUser');
         $cm->mapField(array('fieldName' => 'name'));
 
-        $this->setExpectedException('Doctrine\ODM\MongoDB\MongoDBException');
+        $this->setExpectedException('Doctrine\ODM\MongoDB\Mapping\MappingException');
         $cm->setDiscriminatorField(array('name' => 'name'));
     }
 
@@ -163,7 +163,7 @@ class ClassMetadataTest extends \Doctrine\ODM\MongoDB\Tests\BaseTest
         $cm = new ClassMetadata('Documents\CmsUser');
         $cm->setDiscriminatorField(array('name' => 'name'));
 
-        $this->setExpectedException('Doctrine\ODM\MongoDB\MongoDBException');
+        $this->setExpectedException('Doctrine\ODM\MongoDB\Mapping\MappingException');
         $cm->mapField(array('fieldName' => 'name'));
     }
 
@@ -172,7 +172,7 @@ class ClassMetadataTest extends \Doctrine\ODM\MongoDB\Tests\BaseTest
         $cm = new ClassMetadata('Documents\CmsUser');
         $cm->mapField(array('fieldName' => 'name'));
 
-        $this->setExpectedException('Doctrine\ODM\MongoDB\MongoDBException');
+        $this->setExpectedException('Doctrine\ODM\MongoDB\Mapping\MappingException');
         $cm->mapOneEmbedded(array('fieldName' => 'name', 'targetDocument' => 'CmsUser'));
     }
 
@@ -181,7 +181,7 @@ class ClassMetadataTest extends \Doctrine\ODM\MongoDB\Tests\BaseTest
         $cm = new ClassMetadata('Documents\CmsUser');
         $cm->mapOneEmbedded(array('fieldName' => 'name', 'targetDocument' => 'CmsUser'));
 
-        $this->setExpectedException('Doctrine\ODM\MongoDB\MongoDBException');
+        $this->setExpectedException('Doctrine\ODM\MongoDB\Mapping\MappingException');
         $cm->mapField(array('fieldName' => 'name', 'columnName' => 'name'));
     }
 }
